Refactor ChatMessage.vue by dividing it to ChatMessageImage and ChatMessageText components
